Refactor app_helpers into focused modules
