+++ tests/conftest.py
@@ -135,6 +135,7 @@ def _patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app
     import app_helpers
     import event_db
     import lore_db
+    import migrations
     import state_db
     import story_io
 
@@ -148,7 +149,7 @@ def _patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app
         ("STORIES_REGISTRY_PATH", str(data_dir / "stories.json")),
         ("_LLM_CONFIG_PATH", str(tmp_path / "llm_config.json")),
     )
-    for mod in (app_target, app_helpers, story_io, event_db, lore_db, state_db):
+    for mod in (app_target, app_helpers, story_io, migrations, event_db, lore_db, state_db):
         for attr, value in patches:
             if hasattr(mod, attr):
                 monkeypatch.setattr(mod, attr, value)
